Allow arg_parser to pass '-' as a valid argument
Currently, the arg_parser will say that an empty '-' by itself is an
invalid flag. This is inconvient as '-' without any flags is normally
used to represent stdin or stdout on the comment line. This fix causes
the arg_parser to pass a single '-' along as an extra argument (Like a
filename is) instead of throw an error.